+++ Docs/docs/api/index.md
@@ -1,0 +1,422 @@
+# WindGym API Reference
+
+This is the auto-generated API documentation for the WindGym package.
+
+# API Reference
+
+* [WindGym Package](modules/windgym.md)
+  * [Main Environment Classes](modules/windgym.md#main-environment-classes)
+    * [WindFarmEnv](modules/windgym.md#windfarmenv)
+      * [`WindFarmEnv`](modules/windgym.md#WindGym.WindFarmEnv)
+    * [WindFarmEnvMulti](modules/windgym.md#windfarmenvmulti)
+      * [`WindFarmEnvMulti`](modules/windgym.md#WindGym.WindFarmEnvMulti)
+    * [FarmEval](modules/windgym.md#farmeval)
+      * [`FarmEval`](modules/windgym.md#WindGym.FarmEval)
+    * [AgentEval](modules/windgym.md#agenteval)
+      * [`AgentEval`](modules/windgym.md#WindGym.AgentEval)
+  * [Functions](modules/windgym.md#functions)
+    * [`AgentEvalFast()`](modules/windgym.md#WindGym.AgentEvalFast)
+* [WindGym.core Module](modules/core.md)
+  * [`RewardCalculator`](modules/core.md#WindGym.core.RewardCalculator)
+    * [`RewardCalculator.__init__()`](modules/core.md#WindGym.core.RewardCalculator.__init__)
+    * [`RewardCalculator.calculate_action_penalty()`](modules/core.md#WindGym.core.RewardCalculator.calculate_action_penalty)
+    * [`RewardCalculator.calculate_power_reward()`](modules/core.md#WindGym.core.RewardCalculator.calculate_power_reward)
+    * [`RewardCalculator.calculate_total_reward()`](modules/core.md#WindGym.core.RewardCalculator.calculate_total_reward)
+  * [`WindManager`](modules/core.md#WindGym.core.WindManager)
+    * [`WindManager.__init__()`](modules/core.md#WindGym.core.WindManager.__init__)
+    * [`WindManager.make_wind_direction_list()`](modules/core.md#WindGym.core.WindManager.make_wind_direction_list)
+    * [`WindManager.sample_conditions()`](modules/core.md#WindGym.core.WindManager.sample_conditions)
+  * [`WindConditions`](modules/core.md#WindGym.core.WindConditions)
+    * [`WindConditions.__init__()`](modules/core.md#WindGym.core.WindConditions.__init__)
+    * [`WindConditions.unpack()`](modules/core.md#WindGym.core.WindConditions.unpack)
+    * [`WindConditions.wind_speed`](modules/core.md#WindGym.core.WindConditions.wind_speed)
+    * [`WindConditions.wind_direction`](modules/core.md#WindGym.core.WindConditions.wind_direction)
+    * [`WindConditions.turbulence_intensity`](modules/core.md#WindGym.core.WindConditions.turbulence_intensity)
+  * [`TurbulenceManager`](modules/core.md#WindGym.core.TurbulenceManager)
+    * [`TurbulenceManager.__init__()`](modules/core.md#WindGym.core.TurbulenceManager.__init__)
+    * [`TurbulenceManager.create_sites()`](modules/core.md#WindGym.core.TurbulenceManager.create_sites)
+  * [`WindFarmRenderer`](modules/core.md#WindGym.core.WindFarmRenderer)
+    * [`WindFarmRenderer.__init__()`](modules/core.md#WindGym.core.WindFarmRenderer.__init__)
[…]

[…]
