### 2026-01-05 bf376a4
add better cleanup
