### 2026-01-05 6ff25a2
Ruff
