### 2026-01-05 c627a61
ruff
