### 2026-01-12 8c63201
freeze to current latest dynamiks version
