+++ assets/paper.md
@@ -37,54 +37,63 @@ https://joss.theoj.org/papers/10.21105/joss.06746 -->
 
 **WindGym** is an open-source Python package for reinforcement-learning (RL) based control of wind farms. It provides both single-agent and multi-agent environments, following the Gymnasium API for centralized controllers and the PettingZoo API for multi-agent settings, enabling drop-in use with mainstream RL frameworks [@gymnasium; @pettingzoo]. WindGym is built on top of DYNAMIKS, a multi-fidelity flow simulation framework, which allows users to seamlessly adjust between computational speed and physical fidelity within a single interface [@dynamiks].
 
-WindGym lowers the barrier to reproducible research and benchmarking within the field of RL for wind farm control by standardizing interfaces and providing built-in examples, reward utilities, and tests. The package is MIT-licensed and comes with documentation, continuous integration, and ready-to-run training pipelines, making it straightforward for researchers to prototype, compare, and share RL-based wind farm control strategies.
+The goal of WindGym is to lower the barrier to reproducible research and benchmarking within the field of RL for wind farm control by standardizing interfaces and providing built-in examples, reward utilities, and tests. The package is MIT-licensed and comes with documentation, continuous integration, and ready-to-run training pipelines, making it straightforward for researchers to prototype, compare, and share RL-based wind farm control strategies.
 
 
 
 # Statement of need
 
-Wind energy is projected to play an increasingly important role in global energy production if the transition towards climate neutrality is to become true [@irena2022weto; @iea2021netzero]. Today, most wind turbines are placed closely together in wind farms to leverage shared infrastructure and reduce land use [@Vondelen2024]. However, this introduces what is known as the 'wake effect'. This occurs when an upstream turbine impedes the incoming flow, resulting in a decrease in wind speed and an increase in turbulence for the downstream turbines. This can result in decreased power output and increased structural loads [@Howland2020]. One way to mitigate this phenomenon is with wake steering, where turbines are intentionally misaligned with the wind to help steer the wake away from downstream turbines [@Annoni2018].
 
-Developing control algorithms for wind farms is not a trivial task. One area that has been gaining increased interest is to use RL to help learn a control strategy based on a simulated wind farm environment [@abkar2023reinforcement; @goccmen2025data]. However, even though interest in this field is increasing, much of the work remains fragmented, with many researchers using custom simulators or failing to publish their code bases.
+Wind energy is projected to play an increasingly important role in global energy production if the transition towards climate neutrality is to be realized [@irena2022weto; @iea2021netzero]. Today, most wind turbines are placed closely together in wind farms to leverage shared infrastructure and reduce land use [@Vondelen2024]. However, this introduces the wake effect, where an upstream turbine impedes the incoming flow, resulting in decreased wind speed and increased turbulence for downstream turbines. This can lead to decreased power output and increased structural loads [@Howland2020]. One way to mitigate this phenomenon is wake steering, where turbines are intentionally misaligned with the wind to help steer the wake away from downstream turbines [@Annoni2018].
 
-There already exist a lot of different options for simulating the behaviour of wind farms. These are typically divided into different levels of detail. For example, PyWake [@pywake] and Floris [@FLORIS] are able to simulate the steady-state flow over a full wind farm in the matter of milliseconds, but do not include the transient evolution and turbulent behavior of the wind. Alternatively, simulators such as FOXES [@foxes] and Floridyn [@floridyn] do include the transient evolution of the flow but do not account for the turbulent fluctuations of the wind. 
 
[…]
