+++ assets/paper.bib
@@ -13,7 +13,15 @@
 }
 
 
-
+@misc{WFCRL,
+      title={WFCRL: A Multi-Agent Reinforcement Learning Benchmark for Wind Farm Control}, 
+      author={Claire Bizon Monroc and Ana Bušić and Donatien Dubuc and Jiamin Zhu},
+      year={2025},
+      eprint={2501.13592},
+      archivePrefix={arXiv},
+      primaryClass={cs.LG},
+      url={https://arxiv.org/abs/2501.13592}, 
+}
 
 
 @article{pettingzoo,
