+++ WindGym/utils/generate_layouts.py
@@ -43,6 +43,81 @@ def generate_square_grid(turbine, nx, ny, xDist, yDist):
     return x_pos, y_pos
 
 
+def generate_right_triangle_grid(
+    turbine,
+    nx: int,
+    ny: int,
+    xDist: float,
+    yDist: float,
+    orientation: str = "lower_left",
+):
+    """
+    Create a right-triangle (right-angle) grid of turbines.
+
+    The triangle is created by stacking columns. Column i (0-indexed) contains
+    min(i+1, ny) turbines (so the first column has 1, the second 2, ... up to ny).
+    Legs of the right triangle lie along the x- and y-axes for the "lower_left"
+    orientation. Other orientations are produced by reflections.
+
+    Parameters
+    ----------
+    turbine : WindTurbine
+        The wind turbine object.
+    nx : int
+        Number of columns along the x-direction (length of one leg).
+    ny : int
+        Maximum number of turbines along the y-direction (length of the other leg).
+    xDist : float
+        Distance between turbines in the x-direction, in rotor diameters.
+    yDist : float
+        Distance between turbines in the y-direction, in rotor diameters.
+    orientation : str
+        One of "lower_left" (default), "lower_right", "upper_left", "upper_right".
+        Controls where the right angle is placed.
+
+    Returns
+    -------
+    np.ndarray, np.ndarray
+        Arrays of x and y positions.
+    """
+    if nx <= 0 or ny <= 0:
+        return np.array([]), np.array([])
+
+    D = turbine.diameter()
+    x_spacing = xDist * D
+    y_spacing = yDist * D
+
+    x_list = []
+    y_list = []
+
+    # Build a triangle with the right angle at the lower-left corner:
+    for i in range(nx):
+        count = min(i + 1, ny)  # number of turbines in this column
+        for j in range(count):
+            x_list.append(i * x_spacing)
+            y_list.append(j * y_spacing)
+
+    x_arr = np.array(x_list, dtype=float)
+    y_arr = np.array(y_list, dtype=float)
+
+    # Reflect according to orientation
+    max_x = (nx - 1) * x_spacing
+    max_y = (ny - 1) * y_spacing
+
+    if orientation == "lower_left":
+        return x_arr, y_arr
+    elif orientation == "lower_right":
+        return max_x - x_arr, y_arr
+    elif orientation == "upper_left":
+        return x_arr, max_y - y_arr
+    elif orientation == "upper_right":
+        return max_x - x_arr, max_y - y_arr
+    else:
+        raise ValueError(
+            "orientation must be one of 'lower_left','lower_right','upper_left','upper_right'"
+        )
+
+
 def generate_circle(n, r, angle_offset=0):
     """
     Generate a circular grid of n points with radius r.
@@ -128,6 +203,45 @@ def generate_staggered_grid(
             y_list.append(y)
 
     return np.array(x_list), np.array(y_list)
+
+
+def generate_line_dots_multiple_thetas(X, spacing, thetas, turbine):
+    """
+    Generate X evenly spaced dots on straight lines for multiple angles.
+    Uses a set to ensure points are unique.
+
+    Returns:
+        xs : np.ndarray of x positions
+        ys : np.ndarray of y positions
+    """
+    unique_points = set()
+
+    D = turbine.diameter()
+    spacing *= D  # Convert spacing to meters
+    for theta in thetas:
+        theta_rad = np.deg2rad(theta)
+
+        # Step along the line direction
+        dx = spacing * np.cos(theta_rad)
+        dy = spacing * np.sin(theta_rad)
+
+        # Center the line
[…]
