+++ WindGym/wrappers/per_turbine_wrapper.py
@@ -1,0 +1,223 @@
+"""
+PerTurbineObservationWrapper Template
+
+This wrapper restructures the flat observation/action spaces of WindFarmEnv
+into per-turbine.
+
+-------------------------------------------------------
+Properties:
+    - n_turbines: int
+    - turbine_positions: np.ndarray of shape (n_turbines, 2)
+    - rotor_diameter: float
+    - mean_wind_direction: float
+
+Methods:
+    - reset() -> obs of shape (n_turbines, obs_dim_per_turbine)
+    - step(action) -> obs of shape (n_turbines, obs_dim_per_turbine)
+
+The action input to step() will be shape (n_turbines, action_dim_per_turbine)
+and needs to be flattened before passing to the base environment.
+"""
+
+import gymnasium as gym
+import numpy as np
+from gymnasium import spaces
+from typing import Tuple, Dict, Any, Optional
+
+
+class PerTurbineObservationWrapper(gym.Wrapper):
+    """
+    Wrapper that restructures WindFarmEnv observations and actions
+    from flat vectors to per-turbine format.
+
+    Base env observation (flat): [t0_ws, t0_wd, t0_yaw, t1_ws, t1_wd, t1_yaw, ...]
+    Wrapped observation: [[t0_ws, t0_wd, t0_yaw], [t1_ws, t1_wd, t1_yaw], ...]
+
+    """
+
+    def __init__(self, env: gym.Env):
+        """
+        Args:
+            env: The base WindFarmEnv (or vectorized version)
+        """
+        super().__init__(env)
+
+        self._n_turbines = env.n_turb
+        self._turbine_positions = np.column_stack([env.x_pos, env.y_pos])
+
+        # Rotor diameter in meters (for position normalization)
+        self._rotor_diameter = env.D
+
+        # Observation dimension per turbine
+        self._obs_dim_per_turbine = len(
+            env.farm_measurements.turb_mes[0].get_measurements()
+        )
+
+        # Action dimension per turbine
+        self._action_dim_per_turbine = 1  # Usually just yaw
+
+        # New observation space: (n_turbines, obs_dim_per_turbine)
+        obs_low = -np.inf  # or get from base env
+        obs_high = np.inf
+        self.observation_space = spaces.Box(
+            low=obs_low,
+            high=obs_high,
+            shape=(self._n_turbines, self._obs_dim_per_turbine),
+            dtype=np.float32,
+        )
+
+    @property
+    def n_turbines(self) -> int:
+        """Number of turbines in the farm."""
+        return self._n_turbines
+
+    @property
+    def turbine_positions(self) -> np.ndarray:
+        """
+        Turbine positions in meters.
+
+        Returns:
+            np.ndarray of shape (n_turbines, 2) with (x, y) coordinates
+        """
+        return self._turbine_positions
+
+    @property
+    def rotor_diameter(self) -> float:
+        """Rotor diameter in meters (used for position normalization)."""
+        return self._rotor_diameter
+
+    @property
+    def mean_wind_direction(self) -> float:
+        """
+        Current mean wind direction in degrees (meteorological convention).
+
+        This is used for transforming positions to wind-relative coordinates.
+        270° means wind comes from the West.
+
+        Returns:
+            float: Wind direction in degrees
+        """
+        return self.env.wd
+
+    # =========================================================================
+    # Core Methods
+    # =========================================================================
+
+    def _reshape_obs_to_per_turbine(self, flat_obs: np.ndarray) -> np.ndarray:
+        """
+        Reshape flat observation to per-turbine format.
+
+        Args:
+            flat_obs: Shape (n_turbines * obs_dim_per_turbine,) or
+                      (batch, n_turbines * obs_dim_per_turbine) for vectorized
+
+        Returns:
[…]
