+++ WindGym/wrappers/__init__.py
@@ -2,3 +2,4 @@ from .record_episode_vals import RecordEpisodeVals
 from .curriculum_wrapper import CurriculumWrapper
 from .power_wrapper import PowerWrapper
 from .adversary_wrapper import AdversaryWrapper
+from .per_turbine_wrapper import PerTurbineObservationWrapper
