### 2026-01-22 #9552 @MarcusBinder
> [@editorialbot](https://github.com/editorialbot) invite [@boisgera](https://github.com/boisgera) as editor
>
> 👋 [@boisgera](https://github.com/boisgera) - are you able to take this one on as editor?
>
> [@MarcusBinder](https://github.com/MarcusBinder) will have to modify the submission to meet the new standards before you are able to begin if so.

@crvernon, it should be okay now. The new sections are added and ready for review 👍
