### 2026-01-22 be3cd6d
Disable caching of turbulence field
