+++ tests/test_basics.py
@@ -1,14 +1,14 @@
-import pytest
-import numpy as np
 from pathlib import Path
-from py_wake.examples.data.hornsrev1 import V80
-from stable_baselines3 import PPO
-from WindGym import WindFarmEnv
-from WindGym import FarmEval, AgentEval, PyWakeAgent
+
+import numpy as np
+import pytest
 from dynamiks.sites.turbulence_fields import MannTurbulenceField
 from gymnasium.utils.env_checker import check_env
-from WindGym import AgentEvalFast
-from WindGym.Agents import RandomAgent, ConstantAgent, BaseAgent
+from py_wake.examples.data.hornsrev1 import V80
+from stable_baselines3 import PPO
+
+from WindGym import AgentEval, AgentEvalFast, FarmEval, PyWakeAgent, WindFarmEnv
+from WindGym.Agents import BaseAgent, ConstantAgent, RandomAgent
 from WindGym.utils.generate_layouts import generate_square_grid
 
 
@@ -72,7 +72,7 @@ def wind_farm_env(turbine, mann_turbulence_field, monkeypatch):
         n_passthrough=0.1,
         config=Path("examples/EnvConfigs/2turb.yaml"),
         turbtype="None",
-        burn_in_passthroughs=0.001,
+        burn_in_passthroughs=0.0001,  # Minimal burn-in for faster tests
     )
 
     yield env
@@ -261,79 +261,34 @@ def test_wind_conditions(wind_farm_env):
     assert len(info["Wind speed at turbines"]) == wind_farm_env.n_turb
 
 
-# def test_extreme_wind_conditions(wind_farm_env):
-#    """Test environment behavior at minimum and maximum wind speeds"""
-#    wind_farm_env.set_wind_vals(ws=wind_farm_env.ws_min, wd=270, ti=0.07)
-#    obs_min, _ = wind_farm_env.reset()
-#    assert not np.any(np.isnan(obs_min))
-#
-#    wind_farm_env.set_wind_vals(ws=wind_farm_env.ws_max, wd=270, ti=0.07)
-#    obs_max, _ = wind_farm_env.reset()
-#    assert not np.any(np.isnan(obs_max))
-
-
 def test_reward_functions(wind_farm_env):
     """Test different reward formulations behave as expected"""
-    wind_farm_env.reset()
+    wind_farm_env.reset(seed=42)  # Use fixed seed for reproducibility
 
-    # Test baseline reward
+    # Test baseline reward returns valid type
     if wind_farm_env.power_reward == "Baseline":
         _, reward, _, _, _ = wind_farm_env.step(
             np.zeros(wind_farm_env.action_space.shape)
         )
         assert isinstance(reward, (np.float32, float))
+        assert np.isfinite(reward), "Reward should be finite"
 
-    # Test action penalty
+    # Test that rewards are computed correctly regardless of action
+    # Note: The config uses action_penalty=0.0, so we can't test penalty effects.
+    # Instead, verify rewards are valid numbers for different actions.
     large_action = np.ones(wind_farm_env.action_space.shape)
     _, reward_large, _, _, _ = wind_farm_env.step(large_action)
+    assert isinstance(reward_large, (np.float32, float))
+    assert np.isfinite(reward_large), "Reward for large action should be finite"
+
     small_action = np.zeros(wind_farm_env.action_space.shape)
     _, reward_small, _, _, _ = wind_farm_env.step(small_action)
-    # assert reward_small > reward_large  # Smaller actions should get less penalty
+    assert isinstance(reward_small, (np.float32, float))
+    assert np.isfinite(reward_small), "Reward for small action should be finite"
 
-
-# def test_measurement_system(wind_farm_env):
[…]
