+++ tests/test_renderer.py
@@ -1,0 +1,307 @@
+"""
+Tests for WindFarmRenderer.
+
+Tests rendering functionality including RGB array generation and
+various rendering modes.
+"""
+
+import pytest
+import numpy as np
+from pathlib import Path
+
+# Use non-interactive backend for testing - must be set before importing pyplot
+import matplotlib
+
+matplotlib.use("Agg")
+import matplotlib.pyplot as plt
+
+from py_wake.examples.data.hornsrev1 import V80
+
+from WindGym import WindFarmEnv
+from WindGym.utils.generate_layouts import generate_square_grid
+from WindGym.core.renderer import WindFarmRenderer
+
+
+@pytest.fixture
+def base_example_data_path():
+    """Provides path to the example configuration directory."""
+    return Path("examples/EnvConfigs")
+
+
+@pytest.fixture
+def renderer_none():
+    """Create a renderer with no render mode."""
+    return WindFarmRenderer(render_mode=None)
+
+
+@pytest.fixture
+def renderer_rgb():
+    """Create a renderer for RGB array output."""
+    return WindFarmRenderer(render_mode="rgb_array")
+
+
+@pytest.fixture
+def renderer_human():
+    """Create a renderer for human viewing."""
+    return WindFarmRenderer(render_mode="human")
+
+
+@pytest.fixture
+def env_with_rendering(base_example_data_path):
+    """Create an environment for rendering tests."""
+    yaml_path = base_example_data_path / Path("Env1.yaml")
+    x_pos, y_pos = generate_square_grid(turbine=V80(), nx=2, ny=2, xDist=5, yDist=5)
+    env = WindFarmEnv(
+        turbine=V80(),
+        x_pos=x_pos,
+        y_pos=y_pos,
+        config=yaml_path,
+        render_mode="rgb_array",
+    )
+    env.reset()
+    yield env
+    env.close()
+
+
+class TestRendererInitialization:
+    """Tests for renderer initialization."""
+
+    def test_renderer_init_none_mode(self, renderer_none):
+        """Test renderer initializes with None mode."""
+        assert renderer_none.render_mode is None
+        assert renderer_none.figure is None
+        assert renderer_none.ax is None
+        assert renderer_none.view is None
+
+    def test_renderer_init_rgb_mode(self, renderer_rgb):
+        """Test renderer initializes with rgb_array mode."""
+        assert renderer_rgb.render_mode == "rgb_array"
+
+    def test_renderer_init_human_mode(self, renderer_human):
+        """Test renderer initializes with human mode."""
+        assert renderer_human.render_mode == "human"
+
+
+class TestRendererRenderMethod:
+    """Tests for the main render method."""
+
+    def test_render_returns_none_for_none_mode(self, env_with_rendering):
+        """Test that render returns None when render_mode is None."""
+        renderer = WindFarmRenderer(render_mode=None)
+        result = renderer.render(
+            env_with_rendering.fs,
+            turbine=env_with_rendering.turbine,
+        )
+        assert result is None
+
+    def test_render_rgb_returns_array(self, env_with_rendering):
+        """Test that rgb_array mode returns a numpy array."""
+        renderer = WindFarmRenderer(render_mode="rgb_array")
+        result = renderer.render(
+            env_with_rendering.fs,
+            turbine=env_with_rendering.turbine,
+        )
+        assert isinstance(result, np.ndarray)
+        assert result.ndim == 3  # Height x Width x Channels
+        assert result.shape[2] == 3  # RGB channels
+
+    def test_render_human_returns_none(self, env_with_rendering):
+        """Test that human mode returns None (displays instead)."""
+        # With Agg backend, plt.show() and plt.pause() are no-ops
+        renderer = WindFarmRenderer(render_mode="human")
+        result = renderer.render(
[…]

[…]
