### 2026-01-26 #9552 @MarcusBinder
Hi @boisgera,

That sounds good! I think it is best to leave it to you as you probably know who would be better for the task then 😄
I personally think that everyone on this search could do a perfectly fine job: [link](https://reviewers.joss.theoj.org/reviewers/search?area_id=&keywords=Reinforcement+learning&language=&name=)

Best,
Marcus
