+++ assets/paper.bib
@@ -174,3 +174,13 @@ journal = {Journal of Physics: Conference Series}
   year={2017}
 }
 
+@inproceedings{Neustroev2022,
+  title     = {Deep Reinforcement Learning for Active Wake Control},
+  author    = {Neustroev, Grigory and Andringa, Sytze P.E. and Verzijlbergh, Remco A. and de~Weerdt, Mathijs M.},
+  booktitle = {International Conference on Autonomous Agents and Multi-Agent Systems},
+  year      = {2022},
+  address   = {Online},
+  publisher = {IFAAMAS},
+  month     = {May},
+  numpages  = {10}
+}
