+++ assets/paper.md
@@ -52,9 +52,7 @@ Developing control algorithms for wind farms is not a trivial task. One area tha
 
 # State of the Field
 
-Several options exist for simulating wind farm behaviour at different fidelity levels. PyWake [@pywake] and FLORIS [@FLORIS] simulate steady-state flow over a full wind farm in milliseconds but do not include transient evolution or turbulent behaviour. FOXES [@foxes] and Floridyn [@floridyn] include transient flow evolution but do not account for turbulent fluctuations. 
-
-When we began developing WindGym, no existing package combined dynamic wind farm simulation with standard RL interfaces. While WFCRL [@WFCRL] has since emerged, providing RL environments built on Fastfarm [@fastfarm] and Floris, WindGym offers a distinct advantage: it is built on DYNAMIKS [@dynamiks], a multi-fidelity framework that allows users to interchange fidelity levels within a single codebase. This means researchers can train agents using fast, low-fidelity simulations and validate them with higher-fidelity models without changing their RL setup. Additionally, WindGym provides both single-agent and multi-agent environments through Gymnasium and PettingZoo APIs, whereas WFCRL currently focuses on multi-agent scenarios.
+When we began developing WindGym, no existing package combined dynamic wind farm simulation with standard RL interfaces. The package *wind-farm-env* [@Neustroev2022] existed as the only existing open source option, but it is built on Floris [@FLORIS], with no obvious way of implementing transient wake behaviour. Since then, *WFCRL* [@WFCRL] has emerged, providing RL environments built on Fastfarm [@fastfarm] and Floris. We belive that WindGym offers a distinct advantage. Because it is built on DYNAMIKS [@dynamiks], a multi-fidelity framework that allows users to interchange fidelity levels within a single codebase, researchers can train agents using fast, low-fidelity simulations and validate them with higher-fidelity models without changing their RL setup. Additionally, WindGym provides both single-agent and multi-agent environments through Gymnasium and PettingZoo APIs, whereas *WFCRL* currently focuses on multi-agent scenarios.
 
 
 # Software Design
