### 2026-02-01 e79da4b
Add info about sensors pr turbine
