### 2026-02-10 #9552 @boisgera
> Hey [@lorenzoschena](https://github.com/lorenzoschena), thanks for taking this on!
>
> Any timeline on this? 😄

Hi @MarcusBinder,

Thanks for the nudge ☺️. Sorry for not giving your more info, that's on me!

  - @lorenzoschena told me when he accepted to review your submission that he needed a couple of weeks to get started.

  - Additionally, we need at least two reviewers to open officially the review and lorenzo was the only positive response I got from my first batch of invites. So I need to find and contact a suitable second batch of potential reviewers; I'm doing it right now.
