### 2026-02-10 #9946 @editorialbot
[REVIEW]: WindGym: A Reinforcement Learning Environment for Wind Farm Control
**Submitting author:** @MarcusBinder (<a href="http://orcid.org/0009-0001-5760-5225">Marcus Binder Nilsen</a>)
**Repository:** https://github.com/DTUWindEnergy/windgym
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
