### 2026-02-10 #9552 @boisgera
> [@editorialbot](https://github.com/editorialbot) generate my checklist

Hi @Manishms18!

Please execute this command in the review issue : https://github.com/openjournals/joss-reviews/issues/9946

(this is the "pre-review issue" and it's now closed)
