### 2026-02-10 #9552 @boisgera
> [@boisgera](https://github.com/boisgera) can you please help me with above issue, I am not able to create checklist

See my comment above. This is arguably a bit confusing!
