### 2026-02-13 #9946 @Manishms18
@boisgera This submission presents WindGym, an open-source reinforcement learning environment for wind farm control that provides standardized RL interfaces and integrates multi-fidelity simulation backends.

The software addresses a clear need in the research community by improving reproducibility and enabling consistent benchmarking of RL-based wind farm control strategies. The repository is publicly available, appropriately licensed, well documented, and includes automated tests, examples, and contribution guidelines. The manuscript clearly describes the software’s purpose, design, and scholarly relevance.

I recommend acceptance pending minor improvements, such as strengthening evidence of external adoption, providing fully reproducible environment specifications, and including a minimal quick-start example to improve onboarding.
