### 2026-02-13 884d7cc
Cleanup install
