+++ assets/paper.bib
@@ -17,6 +17,7 @@
       title={WFCRL: A Multi-Agent Reinforcement Learning Benchmark for Wind Farm Control}, 
       author={Claire Bizon Monroc and Ana Bušić and Donatien Dubuc and Jiamin Zhu},
       year={2025},
+      doi={10.52202/079017-4235},
       eprint={2501.13592},
       archivePrefix={arXiv},
       primaryClass={cs.LG},
@@ -110,7 +111,7 @@
 @article{goccmen2025data,
   title={Data-driven wind farm flow control and challenges towards field implementation: A review},
   author={G{\"o}{\c{c}}men, Tuhfe and Liew, Jaime and Kadoche, Elie and Dimitrov, Nikolay and Riva, Riccardo and Andersen, S{\o}ren Juhl and Lio, Alan WH and Quick, Julian and R{\'e}thor{\'e}, Pierre-Elouan and Dykes, Katherine},
-  doi = {https://doi.org/10.1016/j.rser.2025.115605},
+  doi = {10.1016/j.rser.2025.115605},
   journal={Renewable and Sustainable Energy Reviews},
   volume={216},
   pages={115605},
@@ -121,7 +122,7 @@
 @article{abkar2023reinforcement,
   title={Reinforcement learning for wind-farm flow control: Current state and future actions},
   author={Abkar, Mahdi and Zehtabiyan-Rezaie, Navid and Iosifidis, Alexandros},
-  doi = {https://doi.org/10.1016/j.taml.2023.100475},
+  doi = {10.1016/j.taml.2023.100475},
   journal={Theoretical and Applied Mechanics Letters},
   volume={13},
   number={6},
@@ -156,7 +157,7 @@ journal = {Journal of Physics: Conference Series}
    author = {Jonas Schmidt and Lukas Vollmer and Martin Dörenkämper and Bernhard Stoevesandt}, 
    title = {FOXES: Farm Optimization and eXtended yield Evaluation Software}, 
    doi = {10.21105/joss.05464}, 
-   url = {https://doi.org/10.21105/joss.05464}, 
+   url = {10.21105/joss.05464}, 
    year = {2023}, 
    publisher = {The Open Journal}, 
    volume = {8}, 
