+++ Docs/docs/installation.md
@@ -69,6 +69,7 @@ pip install -e .
 
 Now that you have WindGym installed, you can:
 
+- Follow the [Quick Start](./quick-start.md) guide to create your first environment in a few lines of code
 - Explore the [Examples](https://gitlab.windenergy.dtu.dk/sys/windgym/-/blob/main/examples/README.md) to see WindGym in action
 - Learn about [Core Concepts](./concepts.md) to understand how WindGym works
 - Start with [Example 1](https://gitlab.windenergy.dtu.dk/sys/windgym/-/blob/main/examples/Example%201%20Make%20environment.ipynb) to create your first environment

[…]
