### 2026-02-13 #9946 @MarcusBinder
Hi @Manishms18, thanks for taking this on 😄

In regard to your comments:
- We have just added a 'quick start' pane to the documentation, alongside an example quick start notebook. We also added 3 super simple predefined environments in the presets.py file. These are used in the quickstart. That should help with both onboarding and providing reproducible environment specifications. If you had something else in mind, then please let me know 😄
- We have also changed the pyproject.toml such that it should now work with both pixi and using `pip install -e .` . This should further help any reproducibility/setup
- I am unfortunately not able to strengthen the evidence of external adoption as of yet, as none of the submitted papers have been accepted yet.

I hope that answered your feedback

Best,
Marcus
