### 2026-02-16 #9946 @boisgera
Hi @Manishms18,

Manty thanks for you review! 🙏

I noticed that you did not check the **Data sharing** item, is it an oversight?

Best regards,

Sébastien
