### 2026-03-09 a02406a
Rethink rendering pipeline
