+++ WindGym/core/renderer.py
@@ -9,10 +9,8 @@ from typing import Optional
 import numpy as np
 import matplotlib.pyplot as plt
 from matplotlib.backends.backend_agg import FigureCanvasAgg as FigureCanvas
-from IPython import display
 
 from dynamiks.views import XYView, EastNorthView
-from py_wake.wind_turbines import WindTurbines as WindTurbinesPW
 from dynamiks.utils.geometry import get_east_north_height
 import matplotlib.patheffects as path_effects
 from matplotlib.patches import Ellipse
@@ -30,88 +28,197 @@ class WindFarmRenderer:
     Also provides utility methods for plotting farm layouts and frames.
     """
 
-    def __init__(self, render_mode: Optional[str] = None):
-        """
-        Initialize the renderer.
-
-        Args:
-            render_mode: Rendering mode ("human", "rgb_array", or None)
-        """
+    def __init__(
+        self,
+        render_mode: Optional[str] = None,
+        fix_turbines: bool = False,
+        show_indices: bool = True,
+        fontsize: int = 15,
+        axes_lw: float = 1.0,
+        colorbar_vmax_step: float = 2.0,
+    ):
         self.render_mode = render_mode
+        self.fix_turbines = fix_turbines
+        self.show_indices = show_indices
+        self.fontsize = fontsize
+        self.axes_lw = axes_lw
+        self.colorbar_vmax_step = colorbar_vmax_step
 
         # Rendering objects (initialized lazily)
-        self.figure = None
-        self.ax = None
         self.view = None
         self.a = None  # x-axis linspace
         self.b = None  # y-axis linspace
 
     def init_render(self, fs, turbine):
         """
-        Initialize rendering objects.
-
-        This creates the matplotlib figure, axis, and XYView for rendering.
-        Should be called after the flow simulation is created.
+        Initialize the grid extents and XYView used by all render methods.
 
         Args:
             fs: Flow simulation object
             turbine: Turbine object (for hub_height)
         """
-        plt.ion()
         x_turb, y_turb = fs.windTurbines.positions_xyz[:2]
 
-        self.figure, self.ax = plt.subplots(figsize=(10, 4))
         self.a = np.linspace(-200 + min(x_turb), 1000 + max(x_turb), 250)
         self.b = np.linspace(-200 + min(y_turb), 200 + max(y_turb), 250)
 
-        self.view = XYView(
-            z=turbine.hub_height(), x=self.a, y=self.b, ax=self.ax, adaptive=False
-        )
+        self.view = XYView(z=turbine.hub_height(), x=self.a, y=self.b, adaptive=False)
 
-        plt.close()
+    # ------------------------------------------------------------------
+    # Private helpers
+    # ------------------------------------------------------------------
 
-    def render(self, fs, fs_baseline=None, probes=None, turbine=None):
+    def _ensure_initialized(self, fs, turbine):
+        """Initialize view lazily, raising if turbine is not available."""
+        if self.view is None:
+            if turbine is None:
+                raise RuntimeError(
+                    "Renderer not initialized and turbine not provided. "
+                    "Call init_render() first or pass turbine."
+                )
+            self.init_render(fs, turbine)
+
+    @staticmethod
+    def _resolve_fs(fs, fs_baseline, baseline):
+        """Return the appropriate flow simulation based on the baseline flag."""
+        return fs_baseline if baseline else fs
+
+    @staticmethod
+    def _draw_turbines(
+        ax, x_turb, y_turb, R, yaw_plot, tilt, fontsize=15, show_indices=True
+    ):
+        """Draw turbines as ellipses on ax, with optional index labels."""
+        for ii, (x_, y_, r, yaw_, tilt_) in enumerate(
+            zip(x_turb, y_turb, R, yaw_plot, tilt)
+        ):
+            ax.add_artist(
[…]
