+++ examples/EnvConfigs/Env1.yaml
@@ -13,7 +13,7 @@ farm:
   yaw_min: -45
   yaw_max: 45
   xDist: 4
-  yDist: 4
+  yDist: 3
   nx: 2
   ny: 2
 

[…]
