### 2026-03-23 #9946 @MarcusBinder
Hello @boisgera, have you heard anything? We would love to get this through soon-ish, so we can cite the Joss DOI 😄

Best,

Marcus
