### 2026-03-23 #9946 @boisgera
> Hello [@boisgera](https://github.com/boisgera), have you heard anything? We would love to get this through soon-ish, so we can cite the Joss DOI 😄
>
> Best,
>
> Marcus

Hi @MarcusBinder,

I sent an e-mail to Lorenzo today to ask him when he was planning to start the review ~~, but at this stage I don't have more info.~~

**Update.** Lorenzo confirmed me that he will start the review this week.

Best regards,

Sébastien
