### 2026-03-27 #9946 @lorenzoschena
Dear @MarcusBinder, thank you for this submission. WindGym is a timely and welcome contribution to the RL-for-wind-energy landscape — and I look forward to play with it myself!  I have completed my review and propose minor revisions to strengthen the paper, documentation, and repository ahead of publication. Detailed comments are posted in the repository issue.

Great work!
(@boisgera @Manishms18)
