### 2026-03-27 #9946 @boisgera
The WindGym issue @lorenzoschena refers to is https://github.com/DTUWindEnergy/WindGym/issues/4
