+++ assets/paper.bib
@@ -1,8 +1,10 @@
-@misc{dynamiks,
-  title        = {{DYNAMIKS}: Dynamic Wind System Simulator},
-  author       = {DTU},
-  note         = {\url{https://gitlab.windenergy.dtu.dk/DYNAMIKS/dynamiks}},
-  year         = {2023}
+@article{dynamiks,
+        title={Dynamiks 0.0.4: An open-source Dynamic Wind System Simulator},
+        author={Mads M. Pedersen and Julia Steiner and Marcus Binder Nilsen and Jonas Lohmann and Emily Louise Hodgson and Riccardo Riva and Niels Troldborg and S{\o}ren Juhl Andersen and Gunner Larsen and David Robert Verelst and Pierre-Elouan R\'{e}thor\'{e}},
+        url="https://gitlab.windenergy.dtu.dk/DYNAMIKS/dynamiks",
+        publisher={DTU Wind, Technical University of Denmark},
+        year={2026},
+        month={1}
 }
 
 @article{pywake,
