### 2026-03-31 06dc982
update measurement error examples
