+++ examples/curriculum.py
@@ -1,576 +1,0 @@
-import gymnasium as gym
-import numpy as np
-import torch
-import torch.nn as nn
-from stable_baselines3 import PPO
-from stable_baselines3.common.vec_env import DummyVecEnv
-from stable_baselines3.common.callbacks import CallbackList, BaseCallback
-from stable_baselines3.common.policies import ActorCriticPolicy
-import wandb
-from wandb.integration.sb3 import WandbCallback
-from WindGym import WindFarmEnv
-
-# from WindGym.Simple_Wind_Farm_Env import WindFarmEnv
-from WindGym.Agents import PyWakeAgent
-from py_wake.examples.data.hornsrev1 import V80 as wind_turbine
-import argparse
-
-# Device selection
-device = torch.device(
-    "cuda"
-    if torch.cuda.is_available()
-    else "mps"
-    if torch.backends.mps.is_available()
-    else "cpu"
-)
-print(f"Using device: {device}")
-
-
-class WindFarmMonitor(BaseCallback):
-    def __init__(self, verbose=0):
-        super().__init__(verbose)
-        self.episode_rewards = []
-        self.current_rewards = None
-        self.episode_count = 0
-        self.window_size = 100
-        self.agent_powers = []
-        self.base_powers = []
-
-    # def _on_training_start(self):
-    #    self.window_size = self.training_env.get_attr("lookback_window")[0]
-
-    def _on_step(self) -> bool:
-        if self.current_rewards is None:
-            self.current_rewards = np.zeros(self.training_env.num_envs)
-
-        infos = self.locals["infos"]
-        rewards = self.locals["rewards"]
-        dones = self.locals["dones"]
-
-        self.current_rewards += rewards
-
-        for env_idx, (info, done) in enumerate(zip(infos, dones)):
-            agent_power = info["Power agent"]
-            base_power = info["Power baseline"]
-
-            self.agent_powers.append(agent_power)
-            self.base_powers.append(base_power)
-
-            agent_power_avg = np.mean(self.agent_powers[-self.window_size :])
-            base_power_avg = np.mean(self.base_powers[-self.window_size :])
-            power_ratio = agent_power / base_power if base_power != 0 else 1.0
-
-            wandb.log(
-                {
-                    f"charts/agent_power-{env_idx}": agent_power,
-                    f"charts/base_power-{env_idx}": base_power,
-                    f"charts/agent_power_avg-{env_idx}": agent_power_avg,
-                    f"charts/base_power_avg-{env_idx}": base_power_avg,
-                    f"charts/power_ratio-{env_idx}": power_ratio,
-                    f"charts/step_reward-{env_idx}": rewards[env_idx],
-                    f"global_step-{env_idx}": self.num_timesteps,
-                    f"reward_weight-{env_idx}": info.get("curriculum_weight", 0.0),
-                    f"yaw_diff-{env_idx}": info.get("yaw_diff", 0.0),
-                    f"states/yaw_angle_t1-{env_idx}": info["yaw angles agent"][0],
-                    f"states/yaw_angle_t2-{env_idx}": info["yaw angles agent"][1],
-                    f"states/global_wind_speed-{env_idx}": info["Wind speed Global"],
-                    f"states/global_wind_dir-{env_idx}": info["Wind direction Global"],
-                    f"states/pywake_yaw_t1-{env_idx}": info["pywake_yaws"][0],
-                    f"states/pywake_yaw_t2-{env_idx}": info["pywake_yaws"][1],
-                }
-            )
-
-            if done:
-                self.episode_count += 1
-                self.episode_rewards.append(self.current_rewards[env_idx])
-                recent_rewards = self.episode_rewards[-self.window_size :]
-
-                wandb.log(
-                    {
[…]
