+++ examples/pywake_adversary.py
@@ -1,1001 +1,0 @@
-# docs and experiment results can be found at https://docs.cleanrl.dev/rl-algorithms/ppo/#ppo_continuous_actionpy
-import os
-import random
-import time
-from dataclasses import dataclass, field
-from typing import Callable, Dict, Any
-
-import gymnasium as gym
-import numpy as np
-import torch
-import torch.nn as nn
-import torch.optim as optim
-import tyro
-from torch.distributions.normal import Normal
-from torch.utils.tensorboard import SummaryWriter
-
-# Use specific WindGym imports based on your provided files
-from WindGym import WindFarmEnv
-from WindGym.Measurement_Manager import MeasurementManager, MeasurementType
-from WindGym.Agents.PyWakeAgent import NoisyPyWakeAgent
-from WindGym.utils.generate_layouts import generate_square_grid
-
-# Keep these imports as they were in your original cleanRL script
-from WindGym.wrappers import AdversaryWrapper
-
-# Import turbine type from PyWake
-from py_wake.examples.data.hornsrev1 import V80
-
-import yaml  # For loading temporary YAML
-
-
-# # --- SECTION 1: Argument Parsing and Configuration ---
-@dataclass
-class Args:
-    max_eps: int = 10
-    """the maximum number inflow passes"""
-    exp_name: str = field(
-        default_factory=lambda: os.path.basename(__file__)[: -len(".py")]
-    )
-    """the name of this experiment"""
-    save_interval: int = 100
-    """the interval to save the model NOT USED ATM"""
-    yaml_path: str = (
-        "/work/users/manils/wesc/envs/env40.yaml"  # Assuming a default valid path
-    )
-    """the path to the yaml file"""
-    seed: int = 1
-    """seed of the experiment"""
-    torch_deterministic: bool = True
-    """if toggled, `torch.backends.cudnn.deterministic=False`"""
-    cuda: bool = True
-    """if toggled, cuda will be enabled by default"""
-    track: bool = True
-    """if toggled, this experiment will be tracked with Weights and Biases"""
-    wandb_project_name: str = "RLC_comp"  # Original project name
-    """the wandb's project name"""
-    wandb_entity: str = None
-    """the entity (team) of wandb's project"""
-    save_model: int = 1  # Integer instead of bool. Should (hopefully) work.
-    """whether to save model into the `runs/{run_name}` folder"""
-    turbtype: str = "V80"  # Changed to V80 as used in ParametricAdversarialEnv
-    """the type of the wind turbine"""
-    TI_type: str = "None"  # Retaining original
-    """the type of the turbulence model"""
-    dt_sim: int = 1
-    """the time step of the simulation"""
-    dt_env: int = 1
-    """the time step of the environment"""
-
-    # Adversarial Environment Specific Arguments (from original adversary script)
-    n_passthrough: float = 10.0
-    """Number of flow passthroughs for episode length in the base env"""
-    max_bias_ws: float = 1.0
-    """Max wind speed bias for per-turbine sensors (m/s)."""
-    max_bias_wd: float = 100.0
-    """Max wind direction bias for per-turbine sensors (deg)."""
-
-    # Algorithm specific arguments
-    total_timesteps: int = 1000000
-    """total timesteps of the experiments"""
-    learning_rate: float = 3e-4
-    """the learning rate of the optimizer"""
-    num_envs: int = 1  # Keeping original num_envs from your cleanRL script as 1
-    """the number of parallel game environments"""
-    num_steps: int = 2048
-    """the number of steps to run in each environment per policy rollout"""
-    anneal_lr: int = 0  # I changed the bool to int. Hope its fine.
-    """Toggle learning rate annealing for policy and value networks"""
[…]
