### 2026-03-31 #9946 @MarcusBinder
Thank you! Here it is:
- Version tag: v0.1.0
- Zenodo DOI: https://doi.org/10.5281/zenodo.19353795

Thanks again for the work 😄
