### 2026-04-01 #9946 @crvernon
@MarcusBinder - I just need you to address the following before I accept this for publication:

In your archive:

- [x] Please edit the metadata in your Zenodo archive to make the author order the same as what it is in the paper.  No new release required.

In your paper:

- [x] In your "Research Impact Statement" you state that four research papers currently utilize WindGym.  Could you please provide citations to these as a part of this statement?

Let me know when these have been addressed and I'll accept after confirmation.  Thank you.
