### 2026-04-01 #9946 @MarcusBinder
Hello @crvernon

Thanks! Regarding the 4 papers, they are currently submitted but not yet accepted, so I don’t really know how to reference them. Originally, I had put in the 50 stars as proof of adoption, but I got told to change it.
We would really like to get this joss paper published so we can start citing it in the upcoming work we are about to publish.

How should we handle this problem? I would hate to have to wait for them to be published first, but I guess maybe that is the only way?

Also, I fixed the author order now!

Thanks!
