### 2026-04-07 #9946 @MarcusBinder
>  In your "Research Impact Statement" you state that four research papers currently utilize WindGym. Could you please provide citations to these as a part of this statement?

Hey @crvernon, do you have any idea of how we should approach this? I guess we are in a kind of chicken-and-egg situation 😄
