### 2026-04-09 #9946 @boisgera
Hi @MarcusBinder,

@crvernon has the final say on this, but my suggestions in this context would be:
- either make some preprints of the papers and reference them,
- or add "@unpublished" entries to your bibtex file and use the "note" field to state "Submitted to journal/conference XXX".

> > In your "Research Impact Statement" you state that four research papers currently utilize WindGym. Could you please provide citations to these as a part of this statement?
>
> Hey [@crvernon](https://github.com/crvernon), do you have any idea of how we should approach this? I guess we are in a kind of chicken-and-egg situation 😄
