### 2026-04-22 #9946 @crvernon
@MarcusBinder - I agree with @boisgera :

> or add "@Unpublished" entries to your bibtex file and use the "note" field to state "Submitted to journal/conference XXX".

I believe this qualifies this submission considering the docs statement [here](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain):
>Research impact statement: Evidence of realized impact (publications, external use, integrations) or credible near-term significance (benchmarks, reproducible materials, community-readiness signals). The evidence should be compelling and specific, not aspirational.

Let me know when this has been done.  Thank you.
