+++ assets/paper.bib
@@ -187,3 +187,47 @@ journal = {Journal of Physics: Conference Series}
   month     = {May},
   numpages  = {10}
 }
+
+@misc{adversequick,
+      title={Adversarial Sensor Errors for Safe and Robust Wind Turbine Fleet Control}, 
+      author={Julian Quick and Marcus Binder Nilsen and Andreas Bechmann and Tran Nguyen Le and Pierre-Elouan Mikael Rethore},
+      year={2026},
+      eprint={2604.08750},
+      archivePrefix={arXiv},
+      primaryClass={cs.LG},
+      url={https://arxiv.org/abs/2604.08750},
+      note={Submitted to Journal of Physics: Conference Series (Torque 2026). This is the Accepted Manuscript version of an article accepted for publication in Journal of Physics: Conference Series. IOP Publishing Ltd is not responsible for any errors or omissions in this version of the manuscript or any version derived from it. This Accepted Manuscript is published under a CC BY licence.},
+}
+
+@misc{loadsteo,
+      title={Load constrained wind farm flow control through multi-objective multi-agent reinforcement learning}, 
+      author={Teodor Åstrand and Marcus Binder Nilsen and Iasonas Tsaklis and Tuhfe Göçmen and Pierre-Elouan Réthoré and Nikolay Dimitrov},
+      year={2026},
+      eprint={2604.22795},
+      archivePrefix={arXiv},
+      primaryClass={eess.SY},
+      url={https://arxiv.org/abs/2604.22795},
+      note={Submitted to Journal of Physics: Conference Series (Torque 2026). This is the Accepted Manuscript version of an article accepted for publication in Journal of Physics: Conference Series. IOP Publishing Ltd is not responsible for any errors or omissions in this version of the manuscript or any version derived from it. This Accepted Manuscript is published under a CC BY licence.},
+}
+
+@misc{paper3,
+      title={Accelerating Reinforcement Learning for Wind Farm Control via Expert Demonstrations}, 
+      author={Marcus Binder Nilsen and Julian Quick and Tuhfe Göçmen and Nikolay Dimitrov and Pierre-Elouan Réthoré},
+      year={2026},
+      eprint={2604.22794},
+      archivePrefix={arXiv},
+      primaryClass={eess.SY},
+      url={https://arxiv.org/abs/2604.22794},
+      note={Submitted to Journal of Physics: Conference Series (Torque 2026). This is the Accepted Manuscript version of an article accepted for publication in Journal of Physics: Conference Series. IOP Publishing Ltd is not responsible for any errors or omissions in this version of the manuscript or any version derived from it. This Accepted Manuscript is published under a CC BY licence.},
+}
+
+@misc{paper4,
+      title={Hierarchical RL-MPC Control for Dynamic Wake Steering in Wind Farms}, 
+      author={Marcus Binder Nilsen and Teodor Olof Benedict Åstrand and Tuhfe Göçmen and Pierre-Elouan Réthoré},
+      year={2026},
+      eprint={2604.22797},
+      archivePrefix={arXiv},
+      primaryClass={eess.SY},
+      url={https://arxiv.org/abs/2604.22797},
+      note={This work has been submitted to IFAC for possible publication.},
+}
